docs: 全面更新项目文档体系 + 重构后端领域层与仓库层
- docs: 更新全部 10 个目录的 AGENTS.md/AGENTS_zh.md，对齐当前代码（Axum 0.8, Extension DI, 11 个前端特性模块, zod-schemas 等）
- feat: 新增 frontend/src/zod-schemas/ 运行时响应验证（6 个 Zod 模式文件）
- feat: 前端所有特性 API 模块集成 validateResponse 运行时验证
- feat: vite.config.ts 新增 manualChunks 分包策略（vendor-react, vendor-antd, vendor-utils）
- refactor(domain): 清理 pipe.rs 中 150+ 行死代码，删除 order.rs 未使用的枚举
- refactor(dto): 所有 10 个 DTO 模块统一 derive + 文档注释
- refactor(error): 错误码整理（删除冗余变体，补充文档注释）
- refactor(repos): 13 个仓库层统一分页查询模式 + 软删除过滤 + 文档注释
- refactor(handlers): 补全所有 handler 的文档注释与路由路径注释
- refactor: 删除 backend/src/response.rs 中未使用的 response 类型
- chore: seed_data.py 小幅修正

+++ backend/src/repositories/user_repo.rs
@@ -131,8 +131,8 @@ impl UserRepo {
                  ORDER BY created_at DESC LIMIT ?2 OFFSET ?3",
             )
             .bind(&like)
-            .bind(page_size)
-            .bind(offset)
+            .bind(page_size as i64)
+            .bind(offset as i64)
             .fetch_all(pool)
             .await?;
 
@@ -150,8 +150,8 @@ impl UserRepo {
                  FROM users WHERE deleted_at IS NULL
                  ORDER BY created_at DESC LIMIT ?1 OFFSET ?2",
             )
-            .bind(page_size)
-            .bind(offset)
+            .bind(page_size as i64)
+            .bind(offset as i64)
             .fetch_all(pool)
             .await?;
